Extract RSS parser on Android into separate class
